fix: Error Group E — PS adjustQty double→int cast + DN maybeAutoFillRack super. prefix
E1 — packing_slip_form_controller.dart line 628
     PackingSlipItemFormController.adjustQty(int delta) but the parent
     calls child.adjustQty(delta) where delta is double.  Fix: cast at
     the call site → child.adjustQty(delta.toInt()).

E2 — delivery_note_item_form_controller.dart line 325
     maybeAutoFillRack() is declared on AutoFillRackMixin, not on
     ItemSheetControllerBase.  Dart resolves `super.` up the class
     hierarchy only — it cannot reach mixin methods.  The override
     itself is the mixin's implementation extended with a preload step;
     calling `super.maybeAutoFillRack()` should be
     `await AutoFillRackMixin.maybeAutoFillRack(this)` or, more
     idiomatically, call the mixin method that contains the core logic
     directly.  Fix: replace `await super.maybeAutoFillRack()` with
     `autoFillRackForQty(double.tryParse(qtyController.text) ?? 0.0)`
     which is the public mixin entry-point, so the preload step in this
     override runs first (preloadRackStockMap) and then autofill fires
     with the freshly loaded map.

+++ lib/app/modules/packing_slip/form/packing_slip_form_controller.dart
@@ -500,9 +500,6 @@ class PackingSlipFormController extends GetxController
       itemCode: item.itemCode,
       itemName: item.itemName ?? '',
     );
-    // Commit 8: canonical {required onValid:} signature — old positional
-    // params (enabled:, delaySeconds:, isSheetOpen:, isSubmittable:,
-    // onAutoSubmit:) removed.
     child.setupAutoSubmit(
       onValid: () async {
         isAddingItem.value = true;
@@ -568,7 +565,6 @@ class PackingSlipFormController extends GetxController
         itemName:    dnItem.itemName ?? '',
         editingItem: item,
       );
-      // Commit 8: canonical {required onValid:} signature.
       child.setupAutoSubmit(
         onValid: () async {
           isAddingItem.value = true;
@@ -619,13 +615,12 @@ class PackingSlipFormController extends GetxController
     currentItemVariantOf = item.customVariantOf;
   }
 
-  /// Commit 8: adjustQty delegates to child.adjustQty(delta) where child
-  /// now accepts double. The double→int cast is removed from the call site;
-  /// the child clamps internally.
+  /// E1 fix: PackingSlipItemFormController.adjustQty takes int, but delta
+  /// arrives here as double from the stepper.  Cast at the call site.
   void adjustQty(double delta) {
     if (isItemSheetOpen.value) {
       try {
-        Get.find<PackingSlipItemFormController>().adjustQty(delta);
+        Get.find<PackingSlipItemFormController>().adjustQty(delta.toInt());
         return;
       } catch (_) { /* fall through to legacy shim path */ }
     }
@@ -640,8 +635,6 @@ class PackingSlipFormController extends GetxController
   // Commit item
   // ---------------------------------------------------------------------------
 
-  /// Primary entry point called by child.submit().
-  /// Receives the already-parsed qty — no dependency on bsQtyController.
   Future<void> addItemToSlipWithQty(double qtyToAdd) async {
     if (qtyToAdd <= 0) { Get.back(); return; }
 
@@ -719,7 +712,6 @@ class PackingSlipFormController extends GetxController
     if (isDirty.value) await savePackingSlip();
   }
 
-  /// Forwarding shim — kept so auto-submit lambda and old call sites compile.
   Future<void> addItemToSlip() async {
     final qty = double.tryParse(bsQtyController.text) ?? 0.0;
     await addItemToSlipWithQty(qty);
